scripts/asset_management/requirements.txt
# Asset Management System Dependencies

# HTTP client for R2 API interactions
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Progress bars for batch operations
tqdm>=4.65.0
